+++ sushii-2/src/commands/lastfm/mod.rs
@@ -2,12 +2,12 @@ use serenity::framework::standard::macros::group;
 
 mod loved;
 mod np;
+mod profile;
 mod recent;
 mod set;
 mod topartists;
-mod profile;
 
-use self::{loved::*, np::*, recent::*, set::*, topartists::*, profile::*};
+use self::{loved::*, np::*, profile::*, recent::*, set::*, topartists::*};
 
 #[group]
 #[prefix("fm")]