+++ sushii-2/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+-  `fm profile` command
+-  `userinfo` command
+
 ## [0.1.18] - 2021-01-05
 
 ### Fixed